% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbiome}
\alias{microbiome}
\title{Human Microbiome data}
\format{A data set of 100 samples. The first column \code{Body.fluid} is a vector of class labels
for 5 different body-fluids/sites.
The remaining columns is a predictor matrix \code{X} with 1746 predictor variables.}
\usage{
data(microbiome)
}
\description{
Example data set from the \href{Human Microbiome Project}{http://hmpdacc.org/}.
}
\details{
The vector \code{Body.fluid} is a factor with 5 levels, indicating the
body-fluid/site of each sample. The remaining columns can be used as a predictor matrix. Each row
corresponds to a taxonomic profile from the Human Microbiome Project. Each column contains
readcounts for some genus. Each sample (row) has been normalized to relative values (all rows
sum to 1.0)
}
\examples{
data(microbiome)
y <- microbiome[,1]
X <- as.matrix(microbiome[,-1])

}
\author{
Lars Snipen.
}
\seealso{
\code{\link{poems}}.
}

