% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpda.R
\name{mpda}
\alias{mpda}
\title{Multi-level pattern recognition with PLS+LDA}
\usage{
mpda(y, X, reg = 0.5, prior = NULL, selected = NULL, max.dim = NULL,
  n.seg = 10, verbose = TRUE)
}
\arguments{
\item{y}{Vector of responses, must be a factor with 3 or more levels.}

\item{X}{Numeric matrix of predictor values.}

\item{reg}{The regularization parameter used by \code{\link{pdaDim}}, see below.}

\item{prior}{Vector of prior probabilities, one value for each factor level in \code{y}.}

\item{selected}{Matrix of logicals, indicating selected predictor variables, see below.}

\item{max.dim}{Integer, the maximum number of dimensions to consider in PLS.}

\item{n.seg}{Integer, the number of cross-validation segments in \code{\link{pdaDim}}.}

\item{verbose}{Logical, turns on/off output during computations.}
}
\value{
An \code{mpda} object, which is a list of the pairwise \code{\link{pda}} objects. In addition,
the \code{mpda} object has two attributes: A copy of the full predictor matrix \code{X}
(\code{attr(mpda.obj,"X")})and the factor levels of \code{y} (\code{attr(mpda.obj,"Levels")}).
}
\description{
A multi-level classification method that fits one \code{\link{pda}}-model to
every pair of factor levels.
}
\details{
A multi-level problem means the response \code{y} is a factor with at least 3 levels, i.e.
there are at least 3 distinct class labels (texts, integers, etc) in \code{y}. If you have a 2-level problem,
use \code{\link{pda}}.

For each pair of factor levels, a \code{\link{pda}} model is fitted, using the subset of elements in \code{y}
containing these two levels, and the corresponding rows of \code{X} as predictors. If there are L factor levels,
there are N=L*(L-1)/2 such model-pairs.

If the argument \code{reg} is positive (between 0 and 1) it means the \code{\link{pdaDim}} function is used
to estimate the dimensionality of each model. See \code{\link{pdaDim}} for details on this. If \code{reg} is
negative, all models will be fitted with \code{max.dim} dimensions.

The argument \code{prior}, if specified, indicates the prior probability of each factor level. There must be one value for
each factor level, and in the exact same order as the factor levels. By default, all factor levels have equal prior.

The matrix \code{selected} is used to indicate variable selection. For each pair of factor levels, a \code{\link{pda}}
model is fitted, and only a selection of the predictor variables (columns) in \code{X} may be used. In row k of
\code{selected} is a logical vector, one \code{TRUE/FALSE} value for each column in \code{X}. The \code{TRUE} elements
indicate which columns of \code{X} to use (\code{X[,which(selected[k,])]}). PLEASE NOTE: The rows of \code{selected}
must be ordered according to the factor level pairs:

Row 1: Factor level 1 versus 2

Row 2: Factor level 1 versus 3

...<all unique pairs of ordered levels>...

Row N: Factor level (L-1) versus L
}
\examples{
data(poems)
y <- poems[,1]
X <- as.matrix(poems[,-1])
mp.trn <- mpda(y,X,prior=c(1,1,1),max.dim=10)

}
\seealso{
\code{\link{predict.mpda}}, \code{\link{pdaDim}}.
}
\author{
Lars Snipen.
}
