% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpda.R
\name{plot.mpda}
\alias{plot.mpda}
\alias{summary.mpda}
\alias{summary.mpda}
\title{Plotting and summary of mpda object}
\usage{
\method{plot}{mpda}(x, y = NULL, col = c("tan3", "slategray3"), pch = 15,
  legend.pos = "bottomright", xlab = "PLS dimension 1",
  ylab = "PLS dimension 2", ...)

\method{summary}{mpda}(object, ...)
}
\arguments{
\item{x}{An \code{mpda} object, see below.}

\item{y}{not used.}

\item{col}{Colors, one for each class-label.}

\item{pch}{Markers, one for each class-label.}

\item{legend.pos}{Position of legend.}

\item{xlab}{Text on x-axis.}

\item{ylab}{Text on y-axis.}

\item{...}{Optional graphical arguments.}

\item{object}{An \code{mpda} object, see below.}
}
\description{
Generic functions for plotting and printing the content of a \code{mpda} object.
}
\details{
An \code{mpda} object contains a fitted \code{\link{mpda}} model.

The \code{plot.mpda} function produces a multi-panel plot of all the pairwise \code{\link{pda}} models,
see \code{\link{plot.pda}} for details

The \code{summary.mpda} function will display a text giving the number of samples from each class,
the number of PLS-dimensions used, and the priors used in the model fitting.
}
\examples{
data(poems)
y <- poems[,1]
X <- as.matrix(poems[,-1])
mp.trn <- mpda(y,X,prior=c(1,1,1),max.dim=10)
plot(mp.trn)
summary(mp.trn)

}
\seealso{
\code{\link{mpda}}.
}
\author{
Lars Snipen.
}
