% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{plot.pda}
\alias{plot.pda}
\alias{summary.pda}
\title{Plotting and summary of pda object}
\usage{
\method{plot}{pda}(
  x,
  y = NULL,
  col = c("tan3", "slategray3"),
  pch = c(15, 15),
  legend.pos = "topright",
  xlab = "PLS dimension 1",
  ylab = "PLS dimension 2",
  ...
)

\method{summary}{pda}(object, ...)
}
\arguments{
\item{x}{A \code{pda} object, see below.}

\item{y}{not used.}

\item{col}{Two colors, one for each class-label.}

\item{pch}{Two markers, one for each class-label.}

\item{legend.pos}{Position of legend.}

\item{xlab}{Text on x-axis.}

\item{ylab}{Text on y-axis.}

\item{...}{Optional graphical arguments.}

\item{object}{A \code{pda} object, see below.}
}
\description{
Generic functions for plotting and printing the content of a \code{pda} object.
}
\details{
A \code{pda} object contains a fitted \code{\link{pda}} model.

The \code{plot.pda} function display the samples as markers in the first 2 dimensions of the PLS-score space, and color
the markers by the class label information. If the PLS-score-space has only 1 dimension, the second axis in the plot
is non-informative. Since a \code{\link{pda}} object always has only 2 classes, you always specify a pair of colors
and markers in \code{col} and \code{pch}. The \code{legend.pos} must be one of the texts "bottomright", "bottom",
"bottomleft", "left", "topleft", "top", "topright", "right" and "center", specifying the position of the legend
inside the plot.

The \code{summary.pda} function will display a text giving the number of samples from each class,
the number of PLS-dimensions used, and the priors used in the model fitting.
}
\examples{
data(microbiome)
y <- microbiome[1:40,1]
X <- as.matrix(microbiome[1:40,-1])
p.trn <- pda(y,X,prior=c(0.5,0.5),max.dim=2)
plot(p.trn)
summary(p.trn)

}
\seealso{
\code{\link{pda}}.
}
\author{
Lars Snipen.
}
