% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poems}
\alias{poems}
\title{Poetry data}
\format{A data set of 28 samples. The first column \code{Author} is a vector of class labels
and the remaining columns is a predictor matrix \code{X} with 30 predictor variables.}
\usage{
data(poems)
}
\description{
Example data set containing a response vector \code{Author} and a predictor
matrix with character frequencies for 28 english poems.
}
\details{
The vector \code{Author} is a factor with 3 levels, Blake, Eliot and Shakespeare, indicating the
author of each poem (sample). The remaining columns can be used as a predictor matrix. Each row
contains the relative character frequencies of 30 characters counted in each poem.
}
\examples{
data(poems)
y <- poems[,1]
X <- as.matrix(poems[,-1])

}
\author{
Lars Snipen.
}
\seealso{
\code{\link{poems}}.
}

