% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpda.R
\name{predict.mpda}
\alias{predict.mpda}
\title{Classify based on an mpda object}
\usage{
\method{predict}{mpda}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{\link{mpda}} model.}

\item{newdata}{Matrix of predictor values.}

\item{...}{Additional arguments to \code{\link{lda}}.}
}
\value{
A list with two elements, \code{Classifications} and \code{Post.means}.

The vector \code{Classifications} are the suggested factor level for each sample in \code{newdata}.

The matrix \code{Post.means} contains the posterior mean values for each factor level. For L factor levels,
each level is competing' against the other in (L-1) 'matches'. Each match results in posterior probabilities for the
two competing levels. The \code{Post.mean} scores of level k is simply the average of these (L-1) posterior probabilities.
If a sample has a \code{Post.means} value close to 1 for level k, it means this level is the clear winner in all
pairwise 'matches'.
}
\description{
Classify new data based on a trained \code{mpda} model.
}
\details{
Based on the trained \code{\link{mpda}} model, new data objects (rows of \code{newdata}) are
classified, i.e. assigned to a level of the response factor. Remember that \code{\link{mpda}} does not scale
the predictor matrix, make certain your
\code{newdata} are treated identically to the predictor matrix used to train the \code{\link{mpda}}
model. If no \code{newdata} are available, the training data in the \code{mpda} object are used.

Notice: Only the largest dimension of each pairwise \code{pda} model is used for prediction. When training
the \code{mpda} you would normally use the regularization argument \code{reg} and \code{\link{pdaDim}} to
estimate the dimension for each pariwise model, see \code{\link{mpda}}. Thus, different pairwise \code{pda}
are trained to their optimal dimensionality.
}
\examples{
data(poems)
y <- poems[,1]
X <- as.matrix(poems[,-1])
mp.trn <- mpda(y,X,reg=0.5,prior=c(1,1,1),max.dim=3)
lst <- predict(mp.trn)
print(table(y,lst$Classifications))

}
\author{
Lars Snipen.
}
\seealso{
\code{\link{mpda}}.
}

