% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\name{predict.pda}
\alias{predict.pda}
\title{Classify based on a pda object}
\usage{
\method{predict}{pda}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{\link{pda}} model.}

\item{newdata}{Matrix of predictor values.}

\item{...}{Additional arguments to \code{\link{lda}}.}
}
\value{
A list with one element for each dimension. In each element is another list with the integer
\code{Dimension}, the vector \code{Classifications}, the posterior probabilities \code{Posteriors} and
the PLS-scores \code{Scores} used to make the classifications.
}
\description{
Classify new data based on a trained \code{pda} model.
}
\details{
Based on the trained \code{\link{pda}} model, new data objects (rows of \code{newdata}) are
classified. Remember that \code{\link{pda}} does not scale the predictor matrix, make certain your
\code{newdata} are treated identically to the predictor matrix used to train the \code{\link{pda}}
model.
}
\examples{
data(microbiome)
y <- microbiome[1:40,1]
X <- as.matrix(microbiome[1:40,-1])
p.trn <- pda(y[-1],X[-1,],prior=c(0.5,0.5),max.dim=8)    # leaving out sample 1 before training
lst <- predict(p.trn,newdata=X[1,,drop=FALSE])           # predicting sample 1


}
\seealso{
\code{\link{pda}}.
}
\author{
Lars Snipen.
}
