% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{vipCriterion}
\alias{vipCriterion}
\title{The VIP criterion}
\usage{
vipCriterion(pls.model, dim = 1)
}
\arguments{
\item{pls.model}{Object with fitted PLS-model.}

\item{dim}{Integer, the number of dimensions to consider.}
}
\value{
A vector of VIP scores, one for each variable in the predictor matrix of the fitted
PLS model.
}
\description{
Computes the VIP criterion used to rank variable importance.
}
\details{
After the fitting of a PLS-model, some of the original variables will have more
impact than the others on the prediction of the response. The VIP criterion is one way
to quantify this, see Chong&Yun, 2005. This criterion requires a single response regression
problem, which means a two-class classification problem.

A large VIP indicates the corresponding variable is important. A threshold at 1.0 is often
used, variables with VIP above 1.0 are the important ones.
}
\author{
Lars Snipen.
}
\references{
Chong, Il-Gyo & Jun, Chi-Hyuck, 2005, Performance of some variable selection
methods when multicollinearity is present, Chemometrics and Intelligent Laboratory
Systems 78, 103--112.
}
\seealso{
\code{\link{eliminator}}.
}

